Added better keyboard colour UI.
Accidentally reformatted the Javascript file to 4space so I just reformatted the CSS and HTML to that too, least then the whole of the controller app is the same.

Added an SVG image of the razer keyboard, that can then be used to create profiles.
Inside the SVG are the "en-gb" and "en-us" layouts, The layouts can be switched with some javascript.
Uses SnapSVG to do the SVG stuff.

If anyone edits the SVG using inkscape it will get very unhappy and kill most of the key ID's so run svg_editor to fix that. (mostly)
Had to bump the programs window size up to 1000px as for some reason I just couldn't get the SVG to scale to the size of a div.