Add unimplemented multi-device menu.
TODO's are commented in the code. Simply put, this is a conceptual
framework for choosing a device to configure.

Assumes we'll be identifying devices via their serial number.

By default, if there is only one Razer device plugged in to the system,
the "devices" screen should be bypassed entirely, but for demonstration
purposes, this screen will be displayed.

Only the BlackWidow Chroma has an image right now. (Picture taken and
edited a while ago by me - it's not one ripped off the internet.)