Overhaul lighting and brightness controls. Closes #41, #43, #50.
Adds support for the 3 types of FX: main, scroll and logo.

The app now supports new devices like the Razer DeathAdder and other
similar capable devices.

This commit includes other internal changes:
* New dropdown class.
* New function to get profile metadata.
* Tweak how effect strings are retrieved.
* Smoother transitions between device tabs.

This isn't perfect yet, as colours are due for an overhaul too. This may
be apparent when there are multiple lighting effects for a device.