tray: Overhaul UI to be statically generated.
The tray has been unstable for some time due to 'glitches' when the menu
'dynamically' rebuilt itself.

To address this, the UI has been revamped so options are presented per-
device rather than "the current device". This will be a lot more stable.

Closes #103.
Closes #152.

There is now a new option to apply effects/brightness to all connected
devices.