Button icons use SVG tags.
All the SVGs from the ui/img/button/* folder will be loaded into a
BUTTON_SVGS variable sent to the view when the controller initialises.

This way, buttons that have an icon will be <svg> tags so that CSS can
manipulate its colours (pretty!), without needing the view to externally
load them... or be stuck inside a non-interchangable <img> tag.

The exception to this are the larger effect buttons.