Validate (and defaults) for preferences.
Up to now, the set() function had a failsafe default value in case it
was missing. Now, upon initialisation, the data types and a default
value will be ensured.

This is important as the view will be referencing the same data
structure, and expects keys to be present with valid data.