Controller: Strip out WebKitGTK for PyQt5
Things are now moving over to PyQt5 - something that should've happened
a long time ago. This is the third refactor since v0.3.12 for the v1.0.0
release.

Moving to the (Py)Qt5 UI framework allows much faster development of
features. Chunks of time will no longer be spent on 're-inventing UI
elements' which consumed so much time as a 'hybrid' web app.

Data flow has improved, and it will be a lighter desktop application
like it should be.

This enables #7 to be easily accomplished by integrating with the rest
of the desktop, though to maintain the design, a Qt theme will be
produced later to keep the existing look and feel.