backend: Store colours as list
The implementation has gone a bit wonky, so to simplify things, a
parameter or effect should specify the 'last used' colours as a list.

The application will pick up on colour lists to present them as options.
When parameters are used, the 'root' colours key is ignored.

For example, on OpenRazer:
- A static effect requires 1 colour, no parameters.
- A breath effect with 'dual' parameter requires 2 colours.

Fixes a bug spotted in 0b8fbf0a45675cca62fbf298b6200f19a1e75d04
where the random ripple effect did not appear in the Controller (as
'active' for that option was false)