i18n: Drop string IDs, use inline strings
Preferring to store strings "inline" with the logic, as opposed to a
massive set of variables in a file. This also allows backends to name
them however they like.

The locale script will be updated in a later commit to extract them from
all *.ui and *.py files using gettext.