tests: Fix dbus-launch environment (under root/sudo)
For the OpenRazer unit test, to address the error:

dbus.exceptions.DBusException: org.freedesktop.DBus.Error.NotSupported:
Unable to autolaunch a dbus-daemon without a $DISPLAY for X11

We need the OpenRazer daemon to connect to the same DBUS daemon. Since
we can't add plugdev on the CI server to run things under a regular
user, we need to pass all environment variables (-E) to sudo for DBUS to
run.