controller: Partial implementation of effect editor
Effects will now save their column/row count with the effect, so the
program knows the dimensions even if the device/graphic was not present.

From v0.3.12: The mapping SVG is different. Each key/LED is assigned an
"LED" class and an ID in the format: "x0-y0" (where 0 are values), so
any SVGs modified from v0.3.12 must be updated to work.