CI: Skip edge PPA builds for Ubuntu 18.04
Ubuntu 18.04 "bionic" is throwing an error on Launchpad, and a different
one in a local VM. Possibly because of the version of meson.

While 18.04 is intended to be a supported release when the app goes
stable, time would be better spent elsewhere then digging into this
packaging issue at this time.