tray: Show 'auto start' toggle if Controller not installed
Closes #191.

For quite some time, Polychromatic has a config option on whether the
tray applet should automatically start. To support future features, this
is now part of a 'helper' program.

However, if the user installs a split package that only installs the
tray applet, then the ability to set the option is stuck behind the
Controller program that is not installed.

This commit shows an option to toggle auto start if the Controller is
not installed.