openrazer: Add support for 'available_dpi'
Closes #209.
Addresses openrazer/openrazer#427.

For devices that use 'available_dpi' to specify hardcoded DPI values,
the backend will create the combo box instead of Polychromatic's fancier
variable DPI control. In theory, a DPI slider would be incompatible for
that device. Right now, this is just the DeathAdder 3.5G.

The daemon expects DPI values above 0, so passing -1 for DPI Y is
invalid.

This cleans the application logic as the internal 'dpi_single' is
removed.