openrazer: Override DeathStalker Chroma's matrix
Addresses #335.

This device has a matrix of 12x1, but every second LED (2,4,6,8,10,12)
actually blends with the previous LED (1,3,5,7,9,11).

This is pretty confusing for users, since it breaks the expected
behaviour of each LED being its own colour.

This commit makes a virtual 6x1 grid for this device, and will
transparency handle this keyboard's unusual matrix behaviour.

Requires a version of OpenRazer above 3.0.1 for the matrix dimension
changes.