devices: Improve spacing and arrange effect buttons like grid
As opposed to the effect buttons widths being inconsistent. In event of
long text, then the "grid" will lose this desired design. At the moment,
this shouldn't happen ["Spectrum" used a width of 101]

I originally tried some code to set the button width to the widest
button, however while this did the job, this caused flickering due to
being timer-based (2ms). Event-based would be better, but it's tricky.