openrazer: Workaround broken Pulsate via Python library
Fixes #345.
Workaround for openrazer/openrazer#1575.

Since the OpenRazer Python library is broken for the 'bw2013' endpoint,
we'll workaround it by echoing the bytes directly to the driver file,
which is what the daemon would do anyway.

This also adds the missing 'static' effect for these devices. Apart from
the original Tartarus, all these devices with a "bw_set_pulsate" also
have a "bw_set_static" method.

It is intentional that the icon and active state is missing, as a later
commit will refactor the backend into a class that makes this 'quirk'
easier to support.