Refactor backend class
Up to now, data was passed around using carefully constructed
dictionaries. Now they are classes and objects, providing a more
rigid and true object-oriented design for current and future backend
implementations. This improves code navigation and readability.

The old concept of passing an "internal ID" was dropped, as the get/set
functions are grouped together with support for individual refreshing.
This improves performance as data can be cached without having to reload
via get_device().

As this is a major structural change, many areas were refactored,
including tests. These features were temporarily omitted and will
return in a future commit:

- Device Images
- Error checking UI
- Button/Dialog "Options"
- OpenRazer: DPI Sync Button