Remove broken 'Open Mouse Settings' option
Technically this was fixable, since it just needed to read "id" from
"device.form_factor". It's been broken and forgotten about for 2 years!

In the interests of eliminating technical debt, it'll be removed since
KDE 6 no longer opens this mouse page any more. It was originally added
as a shortcut for the user to configure mouse settings (like
acceleration) via their desktop environment.

In future, we could bring it back with better version checks. Or, add a
message to remind the user that they're worth checking out, since we
can't directly read/set options.