Fix the Wayland appId
Wayland compositors use the appId/desktopFileName window property to associate open windows to the correct .desktop launcher. By default Qt sets this to the name of the binary, which in the case of polychromatic was the binary name of Python. Set it to the correct value so that window associations are correct.