i18n: Fix partially translated QSpinBox suffix due to whitespace
"intltool-extract" loses whitespace for spinbox suffixes. I originally
corrected the problem in "create-locales.sh" by replacing the missing
whitespace, but in hindsight, that was more like a hack.

Instead, let the code try again for a whitespace-less version of the
string if it didn't have a translation the first time. This is less work
for existing translations.