Controller: Hide Y axis for Razer Baslisk Essential
To avoid confusing the user, only show the Y axis for known devices that
only have a single, adjustable DPI axis. Like the Baslisk, which always
sets Y to 0.

Not to be confused with devices that only support a fixed set of DPI
values, like the DeathAdder 3.5G, which don't use this control.

Addresses openrazer/openrazer#2132.