Controller: Fix broken header graphic when tiling application
Turns out the label (one of the three that make up the header graphic)
was trying to enforce a minimum width which isn't necessary. This caused
broken looking graphic when tiling the window or snapping to the side.
The minimum width for the window masked this in general use.

Fixes #522.