CI: Move script for creating artifacts
This legacy script is still useful for a "/opt" kind of installation,
but it's only ever used when the CI creates a zip copy from a CI run.

Users are recommended to install the software using a packages so all
the dependencies are installed. Move the script out of the way for
development purposes.