Build: Drop 'sassc' dependency
'sassc' has been deprecated for half a decade:
https://sass-lang.com/blog/libsass-is-deprecated/

We don't use enough of the Sass language to justify its dependency for
pre-processing CSS, so it'll be better long-term to get rid of it.

QSS is based on CSS, with Qt specifics and limitations. Previously,
concatenation was necessary because of incompatibilities with mixing QSS
and SCSS syntax.

Cherry picked from 812c0f49d3d5