Remove dead code for unused effect types
The code isn't used by the end user. It was more of an early
implementation for 2 new effect types (scripted, layered) that haven't
been fully developed.

New effect type(s) will be introduced as part of a larger overhaul of
the codebase. Till then, remove the code that's dead/unused.