Development: Drop 'stable' branch
This was a legacy branch for manual scripts and installations a very
long time ago. It may possibly had been used when there was once a PPA
recipe on Launchpad (this is the case for OpenRazer's stable branch)

This branch isn't necessary, as there is no development purpose for this
branch. Packaging tools these days are good at checking out specific
tags or commits. Some distros roll with the 'master' branch.